[stable only] Unset the default value of MYSQL_GATHER_PERFORMANCE
This patch unsets the default value of MYSQL_GATHER_PERFORMANCE.
That will avoid the installation of ``dbcounter``. In older releases,
this installation is triggering an issue with the version of
"setuptools", that requires a newer version of "packaging" that is
being installed.

Depends-On: https://review.opendev.org/c/zuul/zuul-jobs/+/924815
Change-Id: I765d87c9c6899b19c2df84796bf92c0656a72d7a
Closes-Bug: #2073585
(cherry picked from commit 123ac1b8e86b9da08c680621c10a2499ecff3a8a)