Update glance image size limit
The image size limit for glance using the unified limits is
set to 1000MB (~1GB). This is pretty low given that a volume's
minimum size is 1GB and when uploaded to glance fills out the
whole limit.

The limit issue can also be seen by a recent tempest change[1]
where uploading two volumes failed[2] across various jobs due
to this limit.

We do have a config option in devstack
``GLANCE_LIMIT_IMAGE_SIZE_TOTAL`` but that will need to be
configured for various jobs and a 2GB seems to be a sensible
default which this patch configures.

[1] https://review.opendev.org/c/openstack/tempest/+/938592
[2] Jan 07 23:05:33 devstack-ceph cinder-volume[909965]: ERROR
oslo_messaging.rpc.server cinder.exception.ImageLimitExceeded:
HTTP 413 Request Entity Too Large: The request returned a 413
Request Entity Too Large. This generally means that rate limiting
or a quota threshold was breached.: The response body:: Project
dfe8648c188d46409349eac2c449c0b4 is over a limit for [Resource
image_size_total is over limit of 1000 due to current usage
1024 and delta 0]

Change-Id: I533b7444e5f71275ea3d5c18914e306b1dbbc5cb