Introduce SERVICES_FOR_TEMPEST variable for localrc
This variable overrides the `ENABLED_SERVICES` global variable only for
the `configure_tempest()` function from `lib/tempest`. If the
`SERVICES_FOR_TEMPEST` variable is not defined then `ENABLED_SERVICES`
is used as it had used it before.

This is useful for cases where Tempest is executed from a remote node.

Change-Id: Ic62e48f2f1eb861ec64f51e03353868076cbcc04