Remove leftover from the usage of the removed lib/neutron module
In the patch [1] lib/neutron module was removed completely but it left
call to the non existing currently function 'start_neutron_api' when the
neutron-api service is enabled.
Devstack should start neutron in the same way in case when q-svc or
neutron-api service is enabled and this patch is removing that leftover
call to the 'start_neutron_api' function and make it behave the same way
for both service names.

Additionally this patch adds service "neutron-api" to be checked when
initial networks are going to be created. It is like that as just one of
the services "q-svc" or "neutron-api" is enough to create those initial
networks.

[1] https://review.opendev.org/c/openstack/devstack/+/865014

Related-bug: #2096912
Change-Id: I1287af6a31f60b4e522f0ce3ea525e3336ffd8ba