Fix Swift rsync/replication configuration
Swift rsync and replication services are up and running, however they
fail to replicate any data if needed and if used with more than one
replica.

This patch removes a deprecated option setting[1] and replaces it with
the required setting to use the correct rsync module suffix.

Additionally it removes an outdated subdirectory suffix in the rsyncd
configuration that has been removed as well[2].

Closes-Bug: #2100272

[1] https://opendev.org/openstack/swift/commit/675145ef4a131d548cc1122689732b9b65e5def4
[2] https://opendev.org/openstack/devstack/commit/0e58d22897457831b9dbf02d66a2f29d43803597

Change-Id: I5283405d00883a4dd11b7c001b1bba3776e576b8