Temporary fixes rhel 9 distros
With [1] py39 constraints drop can't run fips jobs on
centos 9-stream or rocky9.
As a workaround can run with PYTHON3_VERSION: 3.11 in affected
jobs.
Until centos 10-stream support is ready[1] we also need
to install libvirt-python and uwsgi from source as uwsgi and
libvirt rpms are bundled with python3.

[1] https://review.opendev.org/c/openstack/requirements/+/948285
[2] https://review.opendev.org/c/openstack/devstack/+/937251

Depends-On: https://review.opendev.org/c/openstack/devstack/+/946763
Depends-On: https://review.opendev.org/c/openstack/devstack/+/948558
Depends-On: https://review.opendev.org/c/openstack/devstack/+/948786
Depends-On: https://review.opendev.org/c/openstack/devstack/+/948797
Related-Bug: #2109591
Change-Id: I80d4c65ba0728c3e4b18738c6b0d539409f19976