Use py38 version of get-pip.py if needed
get-pip.py's minimum version is now py39, so we have to use the
dedicated version of it in case of py38.

NOTE(elod.illes): tempest-full-py3 and tempest-ipv6-only needed to be
set as non-voting to unblock the gate and get this basic devstack fix
able to land.

Change-Id: Ia9bfd74e2dea7cfbeaf04b3913cacd3db9ca7fe9
(cherry picked from commit 9f47219dbe8b5163315280a2073f730cb606d3a3)