Configure 'manager' role in tempest
In this release, nova is implementing the manager role in policy[depends-on],
and Tempest added (depends-on) a new config option to decide if new
defaults are present in testing env. Setting the manager role availability
in Tempest so that test can use manager role user to perform the required
operation in nova.

Depends-On: https://review.opendev.org/c/openstack/nova/+/953063
Depends-On: https://review.opendev.org/c/openstack/tempest/+/953265

Change-Id: I69e32c7de5a63df1c21979f748b77e512068eeec
Signed-off-by: Ghanshyam Mann <gmaan@ghanshyammann.com>