Configure nova 'service' role in tempest
In this release, nova is implementing the service role in
policy[depends-on], and Tempest being branchless needs to
decide if service defaults are present in testing release/
env (Needed-By). Setting the service role availability in
Tempest so that from this release onward, tests can use
service role user to perform the required operation in nova.

Depends-On: https://review.opendev.org/c/openstack/nova/+/957578
Needed-By: https://review.opendev.org/c/openstack/tempest/+/892639

Change-Id: I463cb85f3fcb9f2fdd7aa4a0a5f2ae49782e3fc1
Signed-off-by: Ghanshyam Maan <gmaan@ghanshyammann.com>